// src/app/core/models/settings/designation.ts
export class Designation {
    designationID: number;
    designationName: string;
    designationNameBangla: string;
    isActive: boolean;
    orderNo: number;
    trainingScheduleRankMapID: number;
    trainingScheduleID: number;
    rankID: number;
    departmentID: number | null;
    departmentName: string;
    departmentNameBangla: string;
    departmentDesignationNameBangla: string;
}